Merge version MM to the main branch
Manual merge of the "Merging_work" branch to the main branch.
Replace the old Michael DP version to the last fusion version of Michael Mouthon.
The merge is do manually because github failed to manage the .mlapp files. If you want to check the commit done by Michael Mouthon, look the branch 'Merging_work".