Correction of Epoching module
Correction of a bug when no event is defined.
MichaelDP had set incorrectely the case when to target event was define. The purpose is that it consider the onset as the first TF. It was working only for the first input file and not the following files. This bug is now corrected.